Update admin components and routing
- Updated AdminDashboard.tsx with latest changes
- Removed obsolete Dashboard.tsx file
- Updated admin routes configuration
